# Optimize HTML color validation