I was surprised to see this bottlenecking my app, so I checked Godot's source code and sure enough, it's quite unoptimized.

Below are some GDScript benchmarks of the current master, dev build, 100k runs - I timed various strings against the empty string case.

**--- Invalid colors ---**
2 digits: 0.4ms
2 digits with #: 37.8ms
9 digits: 0.5ms
9 digits with #: 37.5ms
8 digits, bad char on pos 2: 5.9ms
8 digits with #, bad char on pos 2: 50.3ms
8 digits, bad char on pos 8: 24.3ms
8 digits with #, bad char on pos 8: 64.7ms
10000 digits: 0.9ms
10000 digits with #: 288ms

**--- Valid colors ---**
3 digits: 2.4ms
3 digits with #: 42ms
4 digits: 11.6ms
4 digits with #: 52.7ms
6 digits: 17.4ms
6 digits with #: 57.5ms
8 digits: 23.8ms
8 digits with #: 63.6ms

-----

With this PR, all of them took 0.2-0.5ms.