@lawnjelly 2025-04-01
It would be interesting to compare a look up table for this. Perhaps it gets optimized to a lookup table.